fix(cli-sub-agent): execute selected reviewer roster (#1815)
Thread the auto-selected heterogeneous reviewer roster from selection into multi-review execution so execution uses the same tools instead of rebuilding from tier order.

Add regression coverage for range-driven and large-diff-driven auto paths using gemini-cli, antigravity-cli, and codex tier ordering.

Refs #1815.
