+++ crates/cli-sub-agent/src/review_cmd_reviewers.rs
@@ -1,6 +1,7 @@
 use anyhow::Result;
 use tracing::{info, warn};
 
+use crate::cli::ReviewArgs;
 use crate::review_consensus::{build_reviewer_tools, validate_multi_reviewer_tier_pool};
 use csa_config::{GlobalConfig, ProjectConfig};
 use csa_core::types::ToolName;
@@ -10,6 +11,11 @@ const MAX_AUTO_HETEROGENEOUS_REVIEWERS: usize = 3;
 pub(crate) struct AutoReviewerSelection {
     pub(crate) reviewers: usize,
     pub(crate) selected_tools: Vec<ToolName>,
+}
+
+pub(crate) struct EffectiveReviewerSelection {
+    pub(crate) reviewers: usize,
+    pub(crate) selected_tools: Option<Vec<ToolName>>,
 }
 
 pub(crate) struct AutoReviewerRequest<'a> {
@@ -184,7 +190,9 @@ pub(crate) fn resolve_auto_reviewer_selection(
     )
 }
 
-pub(crate) fn resolve_effective_reviewer_count(request: &AutoReviewerRequest<'_>) -> usize {
+pub(crate) fn resolve_effective_reviewer_selection(
+    request: &AutoReviewerRequest<'_>,
+) -> EffectiveReviewerSelection {
     let auto_reviewer_selection = resolve_auto_reviewer_selection(request);
     if let Some(selection) = auto_reviewer_selection {
         let tool_list = selection
@@ -201,14 +209,44 @@ pub(crate) fn resolve_effective_reviewer_count(request: &AutoReviewerRequest<'_>
                 .unwrap_or("no tier name resolved"),
             tool_list
         );
-        selection.reviewers
+        EffectiveReviewerSelection {
+            reviewers: selection.reviewers,
+            selected_tools: Some(selection.selected_tools),
+        }
     } else {
-        request.requested_reviewers
+        EffectiveReviewerSelection {
+            reviewers: request.requested_reviewers,
+            selected_tools: None,
+        }
     }
+}
+
+pub(crate) fn resolve_effective_reviewer_selection_for_args(
+    args: &ReviewArgs,
+    large_diff_auto_escalation: bool,
+    primary_tool: ToolName,
+    resolved_tier_name: Option<&str>,
+    config: Option<&ProjectConfig>,
+    global_config: &GlobalConfig,
+) -> EffectiveReviewerSelection {
+    resolve_effective_reviewer_selection(&AutoReviewerRequest {
+        requested_reviewers: args.requested_reviewers() as usize,
+        explicit_reviewer_count: args.reviewers.is_some(),
+        single: args.single,
+        scope_is_range: args.range.is_some(),
+        large_diff_auto_escalation,
+        explicit_tool: super::prior_rounds::explicit_review_tool(args),
+        explicit_model_spec: args.model_spec.as_deref(),
+        primary_tool,
+        resolved_tier_name,
+        config,
+        global_config,
+    })
 }
 
 pub(crate) fn resolve_multi_reviewer_pool(
     reviewers: usize,
+    selected_reviewer_tools: Option<&[ToolName]>,
     explicit_tool: Option<ToolName>,
     primary_tool: ToolName,
     resolved_tier_name: Option<&str>,
@@ -237,7 +275,9 @@ pub(crate) fn resolve_multi_reviewer_pool(
         }
     }
 
-    let reviewer_tools = if resolved_tier_name.is_some() && explicit_tool.is_none() {
+    let reviewer_tools = if let Some(selected_tools) = selected_reviewer_tools {
+        selected_tools.to_vec()
+    } else if resolved_tier_name.is_some() && explicit_tool.is_none() {
         let pool = build_selected_tool_subset(
             primary_tool,
             &tier_reviewer_tools,
@@ -266,444 +306,5 @@ pub(crate) fn resolve_multi_reviewer_pool(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        AutoReviewerRequest, resolve_auto_reviewer_selection, resolve_effective_reviewer_count,
-        resolve_multi_reviewer_pool,
-    };
-    use crate::run_helpers::TEST_ASSUME_TOOLS_AVAILABLE_ENV;
-    use crate::test_env_lock::TEST_ENV_LOCK;
-    use csa_config::config::TierConfig;
-    use csa_config::{
-        GlobalConfig, ProjectConfig, ProjectMeta, ResourcesConfig, ReviewConfig, TierStrategy,
-        ToolConfig, ToolSelection,
-    };
-    use csa_core::types::ToolName;
-    use std::collections::HashMap;
-    use tokio::sync::OwnedMutexGuard;
-
-    struct ScopedEnvVarRestore {
-        key: &'static str,
-        original: Option<String>,
-    }
-
-    impl ScopedEnvVarRestore {
-        fn set(key: &'static str, value: &str) -> Self {
-            let original = std::env::var(key).ok();
-            // SAFETY: test-scoped env mutation guarded by a process-wide mutex.
-            unsafe { std::env::set_var(key, value) };
-            Self { key, original }
-        }
-    }
-
-    impl Drop for ScopedEnvVarRestore {
-        fn drop(&mut self) {
-            // SAFETY: restoration of test-scoped env mutation.
-            unsafe {
-                match self.original.take() {
-                    Some(value) => std::env::set_var(self.key, value),
-                    None => std::env::remove_var(self.key),
-                }
-            }
-        }
-    }
-
-    fn assume_review_tools_available() -> (OwnedMutexGuard<()>, ScopedEnvVarRestore) {
-        (
-            TEST_ENV_LOCK.clone().blocking_lock_owned(),
-            ScopedEnvVarRestore::set(TEST_ASSUME_TOOLS_AVAILABLE_ENV, "1"),
-        )
-    }
-
-    fn project_config_with_tier(models: &[&str]) -> ProjectConfig {
-        let mut tool_map = HashMap::new();
-        for tool in csa_config::global::all_known_tools() {
-            tool_map.insert(
-                tool.as_str().to_string(),
-                ToolConfig {
-                    enabled: false,
-                    restrictions: None,
-                    suppress_notify: true,
-                    ..Default::default()
-                },
-            );
-        }
-
-        for tool_name in models.iter().filter_map(|model| model.split('/').next()) {
-            tool_map.insert(
-                tool_name.to_string(),
-                ToolConfig {
-                    enabled: true,
-                    restrictions: None,
-                    suppress_notify: true,
-                    ..Default::default()
-                },
-            );
-        }
-
-        let mut tiers = HashMap::new();
-        tiers.insert(
-            "quality".to_string(),
-            TierConfig {
-                description: "Test tier".to_string(),
-                models: models.iter().map(|model| (*model).to_string()).collect(),
-                strategy: TierStrategy::default(),
-                token_budget: None,
-                max_turns: None,
-            },
-        );
-
-        ProjectConfig {
-            schema_version: 1,
-            project: ProjectMeta::default(),
-            resources: ResourcesConfig::default(),
-            acp: Default::default(),
-            tools: tool_map,
-            review: None,
-            debate: None,
-            tiers,
-            tier_mapping: HashMap::new(),
-            aliases: HashMap::new(),
-            tool_aliases: HashMap::new(),
-            preferences: None,
-            github: None,
-            session: Default::default(),
-            memory: Default::default(),
-            hooks: Default::default(),
-            run: Default::default(),
-            execution: Default::default(),
-            session_wait: None,
-            preflight: Default::default(),
-            vcs: Default::default(),
-            filesystem_sandbox: Default::default(),
-        }
-    }
-
-    fn project_config_with_tier_and_review_tool(
-        models: &[&str],
-        tool: ToolSelection,
-    ) -> ProjectConfig {
-        let mut config = project_config_with_tier(models);
-        config.review = Some(ReviewConfig {
-            tool,
-            ..ReviewConfig::default()
-        });
-        config
-    }
-
-    fn distinct_model_family_count(tools: &[ToolName]) -> usize {
-        let mut families = Vec::new();
-        for tool in tools {
-            let family = tool.model_family();
-            if !families.contains(&family) {
-                families.push(family);
-            }
-        }
-        families.len()
-    }
-
-    #[test]
-    fn auto_reviewer_selection_skips_single_tool_tier() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&["codex/openai/gpt-5.4/high"]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: false,
-            scope_is_range: true,
-            large_diff_auto_escalation: false,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert!(selection.is_none());
-    }
-
-    #[test]
-    fn auto_reviewer_selection_uses_all_distinct_tier_tools_up_to_cap() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-            "opencode/openrouter/sonnet/high",
-            "claude-code/anthropic/sonnet/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: false,
-            scope_is_range: true,
-            large_diff_auto_escalation: false,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::GeminiCli,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        })
-        .expect("multi-tool tier should auto-select reviewers");
-
-        assert_eq!(selection.reviewers, 3);
-        assert_eq!(
-            selection.selected_tools,
-            vec![ToolName::GeminiCli, ToolName::Codex, ToolName::Opencode]
-        );
-        assert_eq!(distinct_model_family_count(&selection.selected_tools), 3);
-    }
-
-    #[test]
-    fn large_diff_auto_escalation_selects_heterogeneous_reviewers_for_non_range_scope() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: false,
-            scope_is_range: false,
-            large_diff_auto_escalation: true,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        })
-        .expect("large non-range diff should auto-select heterogeneous reviewers");
-
-        assert!(selection.reviewers >= 2);
-        assert!(distinct_model_family_count(&selection.selected_tools) >= 2);
-        assert_eq!(
-            selection.selected_tools,
-            vec![ToolName::Codex, ToolName::GeminiCli]
-        );
-    }
-
-    #[test]
-    fn auto_range_reviewer_roster_prefers_review_tool_without_filtering_tier() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier_and_review_tool(
-            &[
-                "codex/openai/gpt-5.4/high",
-                "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-            ],
-            ToolSelection::Whitelist(vec!["codex".to_string()]),
-        );
-        let global = GlobalConfig::default();
-
-        let pool = resolve_multi_reviewer_pool(
-            2,
-            None,
-            ToolName::Codex,
-            Some("quality"),
-            Some(&config),
-            &global,
-        )
-        .expect("preferred single-tool roster should resolve");
-
-        assert_eq!(
-            pool.reviewer_tools,
-            vec![ToolName::Codex, ToolName::GeminiCli]
-        );
-        assert!(
-            pool.tier_reviewer_specs
-                .iter()
-                .any(|resolution| resolution.tool == ToolName::GeminiCli)
-        );
-    }
-
-    #[test]
-    fn auto_reviewer_selection_respects_single_flag() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: true,
-            scope_is_range: true,
-            large_diff_auto_escalation: true,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert!(selection.is_none());
-    }
-
-    #[test]
-    fn auto_reviewer_selection_respects_explicit_reviewer_override() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: true,
-            single: false,
-            scope_is_range: true,
-            large_diff_auto_escalation: true,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert!(selection.is_none());
-    }
-
-    #[test]
-    fn large_diff_auto_escalation_respects_explicit_reviewer_count() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-            "opencode/openrouter/sonnet/high",
-        ]);
-        let global = GlobalConfig::default();
-
-        let reviewers = resolve_effective_reviewer_count(&AutoReviewerRequest {
-            requested_reviewers: 2,
-            explicit_reviewer_count: true,
-            single: false,
-            scope_is_range: false,
-            large_diff_auto_escalation: true,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert_eq!(reviewers, 2);
-    }
-
-    #[test]
-    fn auto_reviewer_selection_respects_explicit_model_spec_override() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: false,
-            scope_is_range: true,
-            large_diff_auto_escalation: true,
-            explicit_tool: None,
-            explicit_model_spec: Some("gemini-cli/google/gemini-3.1-pro-preview/xhigh"),
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert!(selection.is_none());
-    }
-
-    #[test]
-    fn auto_reviewer_selection_respects_explicit_tool_override() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: false,
-            scope_is_range: true,
-            large_diff_auto_escalation: true,
-            explicit_tool: Some(ToolName::Codex),
-            explicit_model_spec: None,
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert!(selection.is_none());
-    }
-
-    #[test]
-    fn large_diff_auto_escalation_requires_two_model_families() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-            "antigravity-cli/google/gemini-3.1-pro-preview/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: false,
-            scope_is_range: false,
-            large_diff_auto_escalation: true,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::GeminiCli,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert!(selection.is_none());
-    }
-
-    #[test]
-    fn auto_reviewer_selection_skips_non_range_scope() {
-        let (_env_lock, _available_guard) = assume_review_tools_available();
-        let config = project_config_with_tier(&[
-            "codex/openai/gpt-5.4/high",
-            "gemini-cli/google/gemini-3.1-pro-preview/xhigh",
-        ]);
-        let global = GlobalConfig::default();
-
-        let selection = resolve_auto_reviewer_selection(&AutoReviewerRequest {
-            requested_reviewers: 1,
-            explicit_reviewer_count: false,
-            single: false,
-            scope_is_range: false,
-            large_diff_auto_escalation: false,
-            explicit_tool: None,
-            explicit_model_spec: None,
-            primary_tool: ToolName::Codex,
-            resolved_tier_name: Some("quality"),
-            config: Some(&config),
-            global_config: &global,
-        });
-
-        assert!(selection.is_none());
-    }
-}
+#[path = "review_cmd_reviewers_tests.rs"]
+mod tests;
